src/requirements-recorder.txt
pyserial>=3.5
hidapi>=0.14.0
# Optional for --plot:
matplotlib>=3.7
